requirements.txt
fastapi 
websocket
uvicorn[standard]
soundfile
transformers
torch
torchaudio
lang_trans
python-multipart
python-socketio
librosa 
python-bidi
arabic_reshaper
